FarmerApp/requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
sqlparse==0.4.3
google-api-python-client==2.64.0
google-cloud-translate==3.8.3
pandas==1.5.0
djangorestframework==3.14.0
